feat: migrate DataclipViewer
